requirements.txt
python-telegram-bot==20.8
requests==2.31.0
aiohttp==3.9.3
python-dotenv==1.0.1
